Student register, update and save operations. Model id uniqueness. Remove dependencies.

+++ RUappShared/Model/JSON.swift
@@ -137,7 +137,7 @@ enum JSON {
     }
     
     struct RegisteredStudent: Decodable {
-        var studentId: Int
+        var studentId: Int64
         var institution: Institution
         
         enum CodingKeys: String, CodingKey {
